Allows Ajax content load
The way this is currently working, it's expecting that the contents for
the editor will be provided so that the editor can be loaded with them
when the component is ready in the DOM. In my use case, however, I
needed to load the contents asynchronously. These changes allow for that
by an explicit Vue watcher on the content property. It only communicates
property changes to the editor if they are not equal to what the editor
currently has; a not so elegant way to see where the changes are coming
from: the editor itself or changes to the component's content property
from outside the component.